change: rename files for Folder page
