+++ src/pages/Capacity/Folder.tsx
@@ -1,7 +1,7 @@
 import React, { useEffect, useState } from 'react';
 import { Link, useParams } from 'react-router-dom';
 import UniversimeApi from '@/services/UniversimeApi';
-import './Playlist.css';
+import './Folder.css';
 import NotFoundVideo from './Components/NotFoundVideo/NotFoundVideo';
 import InfoButton from './Components/InfoButton/InfoButton';
 import Footer from '@/components/Footer/Footer';
