+++ src/services/routes.tsx
@@ -13,7 +13,7 @@ import { OAuth2Element }  from './oauth2-google';
 import CapacityPage from "@/pages/Capacity/Capacity";
 import CategoryPage from "@/pages/Capacity/Category";
 import VideoPage from "@/pages/Capacity/VideoPlayer";
-import FolderPage from "@/pages/Capacity/Playlist";
+import FolderPage from "@/pages/Capacity/Folder";
 import ManagerCapacity from "@/pages/Capacity/ManagerCapacity";
 import SignUpPage from "@/pages/SignUp";
 import ManageProfilePage, { ManageProfileLoader } from "@/pages/ManageProfile";
